// hooks/contact/type.ts
export interface ContactGlobal {
  id: number;
  heroTitle?: string | null;
  heroSubtitle?: string | null;
  coordonneesTitle?: string | null;
  coordonneesSubtitle?: string | null;
  address?: string | null;
  phone?: string | null;
  email?: string | null;
  hours?: string | null;
  mapLatitude?: number | null;
  mapLongitude?: number | null;
  createdAt?: string | null;
  updatedAt?: string | null;
}